Fix serializers and views to pass tests
Change how nested relationships work-- just make hyperlinks in note
sections. As in, when a citation section references a document, don't
embed that document-- just link to it. REST, REST, REST.

Refactor both so that the logic of getting a serializer from a section
type isn't duplicated over and over.

Rearrange stuff, etc.