Refactor backbone views & move css/markup out
I tried to remove as much as possible of the presentation/markup out of
the backbone views and into the LESS stylesheets or templates file
(which has warts of its own, but not something I'm going to change right
now [1])

Notably, I removed the AddSection view and just folded it into the
NoteSectionList view, since the distinction wasn't all that clear and
the latter had to stick its tentacles into the former all the time.

[1] Writing HTML in javascript is stupid. No one should ever do it,
really. An idea I've had (that might be not great in its own way) is to
write templates in HTML files and serve them somewhere in the django
project (templates?- but maybe not because they're not actually django
templates). They could be loaded and included in the javascript via a
template tag which could either discover all the different templates
automatically or manually declare their paths. This would have its own
problems, but this is web development and I'm not a pro or a
perfectionist, OK? At least we wouldn't be writing HTML through explicit
string concatenation.