Move backbone templates to HTML files
This creates a new template tag called make_backbone_templates which
scans a given directory for HTML files and converts them to a js script
tag of underscore templates. This way we don't have to make those
templates with string concatenation in javascript. This doesn't yet
solve the problem of defining pretty much identical templates in two
different places (for backbone & django), but it makes things a little
bit more sane for the time being.